+++ src/widget/grid.rs
@@ -119,8 +119,8 @@ impl<M: 'static> Widget<M> for Grid<M> {
     fn paint(&mut self, _ctx: &mut PaintCtx, out: &mut Vec<Instance>) {
         if self.color.a() > 0 {
             out.push(Instance::ui(
-                Position::new(self.x, self.y),
-                Size::new(self.w, self.h),
+                Position::new(self.x as f32, self.y as f32),
+                Size::new(self.w as f32, self.h as f32),
                 self.color,
             ));
         }
